# asset/swagger/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/shops":{"get":{"tags":["shop-controller"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListShopResponse"}}}}}},"post":{"tags":["shop-controller"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShopRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseShopResponse"}}}}}}},"/api/v1/search":{"post":{"tags":["search-controller"],"operationId":"search","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListShopResponse"}}}}}}},"/api/v1/products":{"get":{"tags":["product-controller"],"operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListProductResponse"}}}}}},"post":{"tags":["product-controller"],"operationId":"save_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProductResponse"}}}}}}},"/api/v1/orders":{"get":{"tags":["order-controller"],"operationId":"findAllByMemberId","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListOrderResponse"}}}}}},"post":{"tags":["order-controller"],"operationId":"save_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/admin/api/v1/shops":{"get":{"tags":["admin-shop-controller"],"operationId":"getAllShops","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Shop"}}}}}}},"post":{"tags":["admin-shop-controller"],"operationId":"createShop","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Shop"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Shop"}}}}}}},"/admin/api/v1/products":{"get":{"tags":["admin-product-controller"],"operationId":"getAllProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}},"post":{"tags":["admin-product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/admin/api/v1/orders":{"get":{"tags":["admin-order-controller"],"operationId":"getAllOrders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}}}},"post":{"tags":["admin-order-controller"],"operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}}}}},"/admin/api/v1/members":{"get":{"tags":["admin-member-controller"],"operationId":"getAllMembers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Member"}}}}}}},"post":{"tags":["admin-member-controller"],"operationId":"createMember","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Member"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Member"}}}}}}},"/admin/api/v1/categories":{"get":{"tags":["admin-category-controller"],"operationId":"getAllCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Category"}}}}}}},"post":{"tags":["admin-category-controller"],"operationId":"createCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Category"}}}}}}},"/api/v1/shops/{id}":{"get":{"tags":["shop-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseShopResponse"}}}}}}},"/api/v1/shops/recommend":{"get":{"tags":["shop-controller"],"operationId":"recommend","parameters":[{"name":"count","in":"query","required":false,"schema":{"minimum":1,"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListShopResponse"}}}}}}},"/api/v1/products/{id}":{"get":{"tags":["product-controller"],"operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProductResponse"}}}}}}},"/api/v1/products/shop/{shopId}":{"get":{"tags":["product-controller"],"operationId":"findAllByShopId","parameters":[{"name":"shopId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListProductResponse"}}}}}}},"/api/v1/orders/{id}":{"get":{"tags":["order-controller"],"operationId":"findByIdAndMemberId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOrderResponse"}}}}}}},"/api/v1/categories":{"get":{"tags":["category-controller"],"operationId":"findAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCategoryResponse"}}}}}}},"/api/v1/auth/profile":{"get":{"tags":["auth-controller"],"operationId":"profile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProfileResponse"}}}}}}},"/admin/api/v1/shops/{id}":{"get":{"tags":["admin-shop-controller"],"operationId":"getShopById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Shop"}}}}}}},"/admin/api/v1/shops/exists/{id}":{"get":{"tags":["admin-shop-controller"],"operationId":"shopExists","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/admin/api/v1/products/{id}":{"get":{"tags":["admin-product-controller"],"operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/admin/api/v1/products/shop/{shopId}":{"get":{"tags":["admin-product-controller"],"operationId":"getProductsByShopId","parameters":[{"name":"shopId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/admin/api/v1/products/exists/{id}":{"get":{"tags":["admin-product-controller"],"operationId":"productExists","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/admin/api/v1/orders/{id}":{"get":{"tags":["admin-order-controller"],"operationId":"getOrderById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}}}}},"/admin/api/v1/orders/exists/{id}":{"get":{"tags":["admin-order-controller"],"operationId":"orderExists","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/admin/api/v1/members/{id}":{"get":{"tags":["admin-member-controller"],"operationId":"getMemberById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Member"}}}}}}},"/admin/api/v1/members/exists/{id}":{"get":{"tags":["admin-member-controller"],"operationId":"memberExists","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/admin/api/v1/categories/{id}":{"get":{"tags":["admin-category-controller"],"operationId":"getCategoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Category"}}}}}}},"/admin/api/v1/categories/exists/{id}":{"get":{"tags":["admin-category-controller"],"operationId":"categoryExists","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}}},"components":{"schemas":{"ShopRequest":{"required":["address","category","closeTime","deliveryFee","imageUrl","latitude","longitude","minimumOrder","name","openTime","phone"],"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"category":{"type":"string"},"imageUrl":{"type":"string"},"openTime":{"type":"string"},"closeTime":{"type":"string"},"minimumOrder":{"type":"integer","format":"int64"},"deliveryFee":{"type":"integer","format":"int64"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"ApiResponseShopResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/ShopResponse"},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"ProductResponse":{"required":["category","description","enabled","id","imageUrl","name","price","shopId"],"type":"object","properties":{"shopId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"integer","format":"int64"},"category":{"type":"string"},"imageUrl":{"type":"string"},"enabled":{"type":"boolean"},"id":{"type":"string"}}},"ShopResponse":{"required":["address","category","closeTime","deliveryFee","email","enabled","id","imageUrl","isOpened","latitude","longitude","minimumOrder","name","openTime","phone","products","rating"],"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"imageUrl":{"type":"string"},"isOpened":{"type":"boolean"},"enabled":{"type":"boolean"},"openTime":{"type":"string"},"closeTime":{"type":"string"},"category":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}},"minimumOrder":{"type":"integer","format":"int64"},"deliveryFee":{"type":"integer","format":"int64"},"rating":{"type":"number","format":"double"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"SearchRequest":{"required":["category","name"],"type":"object","properties":{"name":{"type":"string"},"category":{"type":"string"}}},"ApiResponseListShopResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ShopResponse"}},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"ProductRequest":{"required":["category","description","imageUrl","name","price","shopId"],"type":"object","properties":{"shopId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"integer","format":"int64"},"category":{"type":"string"},"imageUrl":{"type":"string"}}},"ApiResponseProductResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/ProductResponse"},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"OrderProductRequest":{"required":["productId","quantity"],"type":"object","properties":{"productId":{"type":"string"},"quantity":{"type":"integer","format":"int64"}}},"OrderRequest":{"required":["products","shopId"],"type":"object","properties":{"shopId":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/OrderProductRequest"}}}},"ApiResponseString":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"string"},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"SignupRequest":{"required":["address","email","isValid","latitude","login","longitude","name","nickname","password","passwordConfirm","phone"],"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"},"passwordConfirm":{"type":"string"},"name":{"type":"string"},"nickname":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"latitude":{"maximum":64,"minimum":1,"type":"number","format":"double"},"longitude":{"maximum":64,"minimum":1,"type":"number","format":"double"},"isValid":{"type":"boolean"}}},"ApiResponseBoolean":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"boolean"},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"LoginRequest":{"required":["login","password"],"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}}},"Shop":{"required":["address","category","closeTime","createdAt","deliveryFee","enabled","id","imageUrl","latitude","longitude","minimumOrder","name","openTime","owner","phone","rating","updatedAt","version","visible"],"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"category":{"type":"string"},"imageUrl":{"type":"string"},"enabled":{"type":"boolean"},"visible":{"type":"boolean"},"openTime":{"type":"string"},"closeTime":{"type":"string"},"rating":{"type":"number","format":"double"},"minimumOrder":{"type":"integer","format":"int64"},"deliveryFee":{"type":"integer","format":"int64"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"owner":{"type":"string"},"id":{"type":"string"},"version":{"type":"integer","format":"int64"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}}},"Product":{"required":["category","createdAt","description","enabled","id","imageUrl","name","price","shopId","updatedAt","version","visible"],"type":"object","properties":{"shopId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"integer","format":"int64"},"category":{"type":"string"},"imageUrl":{"type":"string"},"enabled":{"type":"boolean"},"visible":{"type":"boolean"},"id":{"type":"string"},"version":{"type":"integer","format":"int64"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}}},"Order":{"required":["createdAt","deliveredTime","deliveryAddress","deliveryFee","deliveryPhone","deliveryStatus","grandTotal","id","memberId","paymentMethod","paymentStatus","products","refund","reviewId","shopId","status","updatedAt","version"],"type":"object","properties":{"memberId":{"type":"string"},"shopId":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/OrderProduct"}},"status":{"type":"string","enum":["PENDING","ACCEPTED","CONFIRMED","PREPARING","IN_TRANSIT","COMPLETED","CANCEL_REQUEST","REFUNDED"]},"reviewId":{"type":"string"},"deliveryAddress":{"type":"string"},"deliveryPhone":{"type":"string"},"deliveredTime":{"type":"string"},"deliveryFee":{"type":"integer","format":"int64"},"grandTotal":{"type":"integer","format":"int64"},"paymentMethod":{"type":"string","enum":["EMPTY","CREDIT_CARD","DEBIT_CARD","PAYPAL","MOBILE_PAYMENT","BANK_TRANSFER","CASH_ON_DELIVERY","GIFT_CARD","POINTS","VOUCHER","CRYPTOCURRENCY"]},"paymentStatus":{"type":"string","enum":["PENDING","COMPLETED","FAILED","CANCELED","REFUNDED","PARTIALLY_REFUNDED","DECLINED","IN_PROGRESS","EXPIRED"]},"deliveryStatus":{"type":"string","enum":["PENDING","WAITING_FOR_PICKUP","PICKED_UP","IN_TRANSIT","ARRIVED_AT_DESTINATION","DELIVERED","CANCELLED","FAILED","RETURNED","ON_HOLD"]},"refund":{"$ref":"#/components/schemas/Refund"},"id":{"type":"string"},"version":{"type":"integer","format":"int64"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}}},"OrderProduct":{"required":["price","productId","quantity"],"type":"object","properties":{"productId":{"type":"string"},"quantity":{"type":"integer","format":"int64"},"price":{"type":"integer","format":"int64"}}},"Refund":{"required":["refundAccount","refundAmount","refundBank","refundCompletedDatetime","refundPaymentMethod","refundReason","refundRequestedDatetime","refundStatus"],"type":"object","properties":{"refundStatus":{"type":"string","enum":["PENDING","PARTIALLY_REFUNDED","REFUNDED","REJECTED","CANCELLED","IN_PROGRESS","FAILED","EXPIRED"]},"refundReason":{"type":"string"},"refundRequestedDatetime":{"type":"string"},"refundCompletedDatetime":{"type":"string"},"refundPaymentMethod":{"type":"string","enum":["EMPTY","CREDIT_CARD","DEBIT_CARD","PAYPAL","MOBILE_PAYMENT","BANK_TRANSFER","CASH_ON_DELIVERY","GIFT_CARD","POINTS","VOUCHER","CRYPTOCURRENCY"]},"refundAmount":{"type":"integer","format":"int32"},"refundAccount":{"type":"string"},"refundBank":{"type":"string"}}},"Member":{"required":["address","createdAt","email","id","latitude","login","longitude","name","nickname","password","phone","seller","shopId","updatedAt","version"],"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"nickname":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"latitude":{"type":"string"},"longitude":{"type":"string"},"seller":{"type":"boolean"},"shopId":{"type":"string"},"id":{"type":"string"},"version":{"type":"integer","format":"int64"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}}},"Category":{"required":["createdAt","enabled","id","imageUrl","key","name","order","updatedAt","version"],"type":"object","properties":{"key":{"type":"string"},"name":{"type":"string"},"imageUrl":{"type":"string"},"enabled":{"type":"boolean"},"order":{"type":"integer","format":"int64"},"id":{"type":"string"},"version":{"type":"integer","format":"int64"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}}},"ApiResponseListProductResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"ApiResponseListOrderResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponse"}},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"OrderProductResponse":{"required":["category","description","enabled","id","imageUrl","name","price","quantity","shopId"],"type":"object","properties":{"shopId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"integer","format":"int64"},"category":{"type":"string"},"imageUrl":{"type":"string"},"enabled":{"type":"boolean"},"id":{"type":"string"},"quantity":{"type":"integer","format":"int64"}}},"OrderRefundResponse":{"required":["refundAmount","refundCompletedDatetime","refundPaymentMethod","refundReason","refundRequestedDatetime","refundStatus"],"type":"object","properties":{"refundStatus":{"type":"string","enum":["PENDING","PARTIALLY_REFUNDED","REFUNDED","REJECTED","CANCELLED","IN_PROGRESS","FAILED","EXPIRED"]},"refundReason":{"type":"string"},"refundRequestedDatetime":{"type":"string"},"refundCompletedDatetime":{"type":"string"},"refundPaymentMethod":{"type":"string","enum":["EMPTY","CREDIT_CARD","DEBIT_CARD","PAYPAL","MOBILE_PAYMENT","BANK_TRANSFER","CASH_ON_DELIVERY","GIFT_CARD","POINTS","VOUCHER","CRYPTOCURRENCY"]},"refundAmount":{"type":"integer","format":"int32"}}},"OrderResponse":{"required":["createdAt","deliveredTime","deliveryAddress","deliveryFee","deliveryPhone","deliveryStatus","grandTotal","id","memberId","paymentMethod","paymentStatus","products","refund","reviewId","shop","status"],"type":"object","properties":{"memberId":{"type":"string"},"shop":{"$ref":"#/components/schemas/ShopResponse"},"products":{"type":"array","items":{"$ref":"#/components/schemas/OrderProductResponse"}},"status":{"type":"string","enum":["PENDING","ACCEPTED","CONFIRMED","PREPARING","IN_TRANSIT","COMPLETED","CANCEL_REQUEST","REFUNDED"]},"reviewId":{"type":"string"},"deliveryAddress":{"type":"string"},"deliveryPhone":{"type":"string"},"deliveredTime":{"type":"string"},"deliveryFee":{"type":"integer","format":"int64"},"grandTotal":{"type":"integer","format":"int64"},"paymentMethod":{"type":"string","enum":["EMPTY","CREDIT_CARD","DEBIT_CARD","PAYPAL","MOBILE_PAYMENT","BANK_TRANSFER","CASH_ON_DELIVERY","GIFT_CARD","POINTS","VOUCHER","CRYPTOCURRENCY"]},"paymentStatus":{"type":"string","enum":["PENDING","COMPLETED","FAILED","CANCELED","REFUNDED","PARTIALLY_REFUNDED","DECLINED","IN_PROGRESS","EXPIRED"]},"deliveryStatus":{"type":"string","enum":["PENDING","WAITING_FOR_PICKUP","PICKED_UP","IN_TRANSIT","ARRIVED_AT_DESTINATION","DELIVERED","CANCELLED","FAILED","RETURNED","ON_HOLD"]},"refund":{"$ref":"#/components/schemas/OrderRefundResponse"},"id":{"type":"string"},"createdAt":{"type":"string"}}},"ApiResponseOrderResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/OrderResponse"},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"ApiResponseListCategoryResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponse"}},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"CategoryResponse":{"required":["enabled","imageUrl","key","name","order"],"type":"object","properties":{"key":{"type":"string"},"name":{"type":"string"},"imageUrl":{"type":"string"},"enabled":{"type":"boolean"},"order":{"type":"integer","format":"int64"}}},"ApiResponseProfileResponse":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/ProfileResponse"},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"}}},"ProfileResponse":{"required":["address","email","latitude","login","longitude","name","nickname","phone","seller"],"type":"object","properties":{"login":{"type":"string"},"name":{"type":"string"},"nickname":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"seller":{"type":"boolean"}}}}}}